+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## 变更
 
+* PathSyntaxParser.parse更改返回值类型为tuple[AttrKey | IndexKey, ...]
+* PathSyntaxParser.parse添加缓存
 * 字段TokenInfo.current_token由传入变为自动推导
 * 现在路径语法解析器发现孤立的键元信息会警告"Isolate meta found"
 * 现在路径语法解析器对无效的转义序列有更好的提示信息